Day 4 , get data from www site homepage
You will have to do the following:

Write a program that will read a webpage and puts its text into a file
named "index.xhtml"
The test site will be: http://laroza.freehostia.com/home/index.php
That is the bare minimum requirements, but I would ask you try to make
it work for any site.

You do NOT have to:

Have it do anything but download the single file of a web page, so
ignore any external scripts or css.
Have it to any rendering, just write it to a file (for opening with a
browser, most likely)

For extra credit:

Have either prompt or use a command line argument to do do this with
any webpage
Have it name the output file after the title of the document (hint:
text in the elements <title></title>, you can use xml or html parsers
for this, or you can use regular expressions)

The test page is valid XHTML 1.1 (and is valid XML). All conditions and
extra credit can be done with standard Python. If you use another
language or non standard libs, let me know exactly how to get your
entry working.