Day 5
You will have to do the following:

Write a program that will emulate a common guessing game.
The game should randomly set a number which the user will have to guess
and will get "higher" or "lower" hints as it goes. Only integers will
be used for the target number.
The target number should be 1 to 100 (although this "game" is harldy a
game. Anybody will get it in 7-8 tries without even "guessing", by
using logic)
The game will end with a note of the win when the correct number is
guessed.
The program should tell the number of guesses it took to get it as well.

In addition, these are also part of the challenge:

The code should be well formatted.
The code layout should be structured and logical.
The use of logical design is a big part of this, so keep a consistant
format, and use logical variable and function names.